Update draw functions for separate srcfile
